requirements.txt
langgraph
langchain
langchain-core
pydantic
python-dotenv 
langchain_google_genai
langchain-groq
IPython
langchain_google_vertexai